CDPD-31676 Backport KUDU-3332 to 7.1.7-SP1 (part 1)
[tools] Add a tool to recover master data from tablet servers

This adds a tool that attempts to recover master metadata from
tablet servers using ListTablets, writing the metadata to a
syscatalog table that a new master process can use to make the
cluster operational.

It has several limitations. See the help description for the new tool.

Note that it should be possible to fix some limitations by
giving more master metadata to tablet servers. The advantage
of the tool as-is is that it should work on all versions of
Kudu since 1.0.

Change-Id: If29e421d466a531ebad72e281ae27e74e458f8c6
Reviewed-on: http://gerrit.cloudera.org:8080/9490
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
(cherry picked from commit ab0421b13b81798548b7dc2f7acd8f7362d459c5)
  Conflicts:
    src/kudu/common/partition.h
    src/kudu/tools/kudu-admin-test.cc
    src/kudu/tools/master_rebuilder.cc