CLOUDERA-BUILD Add an option to use CDH Hive/Hadoop
This commit introduces another option to allow the Cloudera
distrubution to be used as the third-party dependencies, especially for
Hive and Hadoop components. It extracts the CDH_VERSION from 'version.txt'.
Then the global build number of the latest distribution in that CDH version
is queried from BuildDB, and the corresponding build is used.

Change-Id: I1bb47a402351765f86db868c4f51a8b4a5b4582a