CLOUDERA-BUILD rename CDH_GBN to avoid naming conflict
The previous commit adds the option to use CDH Hive/Hadoop published
via Impala tarball as the third-party dependecies. However, cauldron
builds Kudu twice, and the first time it's as a precursor to the Impala
build. Since in impala build configuration sript, CDH_GBN is already
defined as the current build number, to avoid naming conflict, this
commit renames it to KUDU_CDH_GBN.

Change-Id: I3ab47fc45f5c6928b254f6a3d9e5f3d40bc8c893