CLOUDERA-BUILD: Use Spark’s spark-avro module
In Spark 2.4 spark-avro is now a part of Spark itself.
This change migrates the Kudu spark-avro
dependencies and adds a test to ensure that
the functionality does not break.

This patch will need to be pushed in sync with the
CDH Spark rebase. It is downstream only for now
given that Spark 2.4 is not released upstream yet.
KUDU-2606 tracks moving this work upstream.

Change-Id: Ie407cccdb579d1818d4207260aae57cec998ff13