CDH-78512: pull Hive, Sentry, Hadoop from Cauldron
This patch restores the behavior from
beb1a4feab41744b062e839e6ba0fef3e1c903dc, wherein prebuilt Hadoop
ecosystem components are pulled from Cloudera's build servers. The major
difference is that this patch hides this behavior with a flag that is
off by default.

In the original patch, there was behavior that wrote the GBN to disk and
checked the local CDH GBN against the latest version in BuildDB. This
seems to have been for the sake of always pulling the latest build upon
subsequent runs of cmake. To simplify this a bit, this patch instead always
pulls from BuildDB when building thirdparty.

To use this, do the following before running the cmake incantation:

  export KUDU_USE_CDH_ECOSYSTEM=1

Change-Id: I30239e0b4e07ac8d6a455ac105e2a77a5c9f752d