CDH-79491: enable KUDU_USE_CDH_ECOSYSTEM by default for test executions
This patch enables the flag which pulls prebuilt Hadoop ecosystem components
from Cauldron by default for all test executions. It also fixes Sentry
related tests failure due to the configuration divergence between CDH
Sentry and Apache Sentry, which certain Kudu actions (e.g. ALTER) are not
permitted to be granted by Sentry clients. The fix is to add all Kudu
actions to this configuration list instead of updating the tests' logic,
because privilege granting should not be part of the test coverage.

Change-Id: Iea4b1955c6dfdde1b1f4c167c9862a5dc1b3c8ef