CLOUDERA_BUILD set dependency versions for local builds
This is the best practice for enabling CDPD local builds: set the various
dependency versions to the "last known good" versions substituted in by a
successful CDPD build. I used commit 94c502721 as a template, which was
authored by a Jenkins job but applied to the wrong branch (cdh6.3.0 instead
of cdpd-master).

Change-Id: I074ada8fad937f379ba55ff3b215e8a4ce536f84