CLOUDERA_BUILD default to KUDU_USE_CDH_ECOSYSTEM=0
When enabled, the thirdparty build gets its Hadoop, Hive, and Sentry
artifacts from the last known good CDH build. Ignoring the fact that
Sentry is gone, this retrieval mechanism is just flat out wrong for CDPD.

We'll probably end up removing this switch, but for the time being let's
just disable it.

Change-Id: I74412d7d1f44ea7ee24be5703e3e2da7fee2e9ab