[CLOUDERA-BUILD] Disable Sentry related tests
This is a downstream only change to disable the Sentry related tests.
This allows us to upgrade to support Hive 3.x, the version in CDP.
The change to upgrade Hive will be in a follow up commit.

Note: This patch is staged upstream so that it can be applied once
upstream can support Hive 3. See here:
https://gerrit.cloudera.org/#/c/14020/

Change-Id: Ie848b93adedae587d5d09a243f40c78c3b701b1b