[CLOUDERA-BUILD] Upgrade Hive dependency to 3.1
This patch upgrades Hive to 3.1.1 and Hadoop to 3.2.0 in Kudu’s thirpdarty
build. Follow on changes should ensure the thirdparty build can use
CDPs versions of Hive and Hadoop via thirdparty.

The Java Hive dependency is also changed to use the CDP version of Hive.

Note: This patch is staged upstream so that it can be applied once
upstream can support Hive 3. See here:
https://gerrit.cloudera.org/#/c/13256/

Change-Id: Ice1319688acbb6c5c20b072042d4b3a78b14b6f2