[hms] Fix TestKuduMetastorePlugin on Hive 4
This patch is another patch in preparation for Hive 3 and 4.
In Hive 4 the MetastoreDefaultTransformer ensures that
external tables don’t use the warehouse as a their storage location.
Kudu doesn’t actually use this location, but the location must still be
valid.

This patch changes the test to set the correct location for external
and managed tables and configures a default external warehouse
for Hive 4+.

Change-Id: Ia211e65f66ce6002e6d25b31ab88437630e19473
Reviewed-on: http://gerrit.cloudera.org:8080/14487
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
(cherry picked from commit d96f8fca55222da2a8a9e4c98bf4bebee3525989)