[nvm_cache] set --nvm_cache_allocation_retry_count to 0
As it turned out, the logic of retrying NVM-based allocations in
NvmLRUCache::AllocateAndRetry() didn't serve the intended purpose.
The actual memory was deallocated only after all the retries, and
the code effectively was attempting to allocate memory again and
again on nothing-yet-freed-cache.

As a stop-gap solution, this patch sets the default value for the
--nvm_cache_allocation_retry_count flag to 0. The matter will be
properly addressed in a separate changelist.

Change-Id: If33d1a7ffdb7389a02c1b86f2f4b3b9dc362b5d9
Reviewed-on: http://gerrit.cloudera.org:8080/14497
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
(cherry picked from commit 24aebbaa77f68c813e604eb157b73e0c9bc7433f)