csd: add commands to enter/exit tserver maintenance
This adds commands to Kudu tservers that get the local UUID and
enter or exit tserver maintenance for that UUID. It uses `kudu fs dump
uuid` to get the UUID and reuses the existing gflag config generator.

CM already has a concept of maintenance mode that simply disables alerts
when hosts go down. To not collide with that, the added commands are
called "Disable Re-replication" and "Re-enable Re-replication".

Testing:
- Deployed this on a real cluster running roughly the 1.11.0 RC2
  candidate.
- "Disable Re-replication" and "Re-enable Re-replication" appear when
  looking at Tablet Server actions [1].
- I used the commands to put a tserver into maintenance and then
  exit maintenance.
- I confirmed that hovering over the commands show the appropriate
  descriptions.
- I also tried putting a tserver that was stopped into maintenance to
  confirm the absence of "requiredRoleState" in the command worked as
  expected. As well, when I exited maintenance mode while the server was
  still down, the cluster started recovering, as expected.
- I also ran ksck through the CM UI after putting one of the tservers
  into maintenance, and in the expanded stdout of the command, I
  confirmed I could see MAINTENANCE_MODE listed.

[1] https://imgur.com/7WtMr1G

Change-Id: I839f23b9c48696752bbe7de79b4cac20188c4150