[csd] Add metrics_url_parameters to CSD
This patch adds a `metrics_url_parameters` parameter to the CSD for
the master and tserver serivce. The default value is `compact=1&level=info`
to reduce the metrics collected by CM without sacrificing the commonly
used and useful metrics.

A follow-on CM change is required to adjust the CM `kudu_adapters.py`
to use the `metrics_url_parameters` if it exists:
https://github.infra.cloudera.com/Starship/cmf/blob/cdpd-master/agents/cmf/src/cmf/monitor/generic/kudu_adapters.py#L124

This patch addresses OPSAPS-41290, CDH-78432.

Change-Id: Id11dc7be19bc23d37a7511b9b89631ec459e4766