csd: enable AutoTLS for Kudu
This patch tells CM that Kudu's web UI participates in automatic TLS
configuration. The effect is that when the CM auto TLS wizard runs, it'll
configure Kudu's web UI parameters appropriately.

As far as I can tell, this is a CM feature that can be applied to any
conformant service. That's why I changed the CSDs for all versions of Kudu
that support TLS in the web UI.

Change-Id: I1c11991bf39c27aaa56071e28cc680e3d6ba1d09