[csd] Add the host to the kudu-monitoring.properties
This patch adds the host to the generated kudu-monitoring.properties
file. A follow on patch in CM will use that hostname, instead of localhost,
when generating the metrics URL if the webserver_interface is not specified.

I tested this manually on the nightly7x-unsecure-1 cluster and verified
the host entries were generated on all of the Kudu instances.

Example entry:
host=nightly7x-unsecure-1.nightly7x-unsecure.root.hwx.site

Change-Id: Iadf75973b865800ed2e7e70c76266446a9ef3708