[csd] time-source specific configuration knobs
This changelist adds Kudu service-wide configuration property to
select among 'auto', 'builtin', and 'system' time sources.  In addition,
it adds a knob to specify the set of servers for the built-in NTP
client in case if 'builtin' time source is selected.  The default setting
for the new control is 'system': this is to make it backward-compatible
with prior versions.

Of course, all the configuration can be done via the flags in the
Advanced Configuration Snippet (Safety Valve) for gflagfile, but
apparently it's worth adding these top-level controls in CM for better
usability.

As for CSD forking, it seems we should be able to get away with adding
the issue with the time source for older releases as 'known issue' and
avoid forking CSD just for this tiny feature.  The older Kudu version
might break (i.e. kudu-master and kudu-tserver fail to start) only if
setting the time source to 'auto'.  All the newly exposed parameter in
this CSD versions do exist in versions of Kudu since the latest fork for
7.1.1, but only the very latest Kudu versions do recognize the 'auto'
time source (by default it's set to 'system' as already mentioned).

I tested this patch at http://aserbin-kudu-c75-5.vpc.cloudera.com:7180

Change-Id: I671c07457e2898c15bfdc24ef966d4cf55f5ecca