csd: CDPD-8367 Add Kudu service creation in Ranger
For Kudu Ranger plugin to function correctly, a Kudu service instance
needs to be created and initialized in Ranger admin server. This patch
achieves it via CM to create and initialize a new Kudu service on demand
in Ranger (if it does not exist), when Kudu starts.

The service naming strategy is the following, so that separate Kudu
services can be created in Ranger without naming conflict for scenarios
that several Kudu clusters sharing the same Ranger admin server:
- Sanitize and use the service name stored in 'ranger.plugin.kudu.service.name',
  which is "cm_kudu" by default.
- If this config is set to "{{GENERATED_RANGER_SERVICE_NAME}}", CM generates
  a sanitized per-service unique name.

The configurations used to create the Ranger Kudu service is defined in
a JSON template in 'aux' folder. To be more specific, sets
'policy.download.auth.users' and 'tag.download.auth.users' so that 'kudu'
user can pull the resource/tag based policies from the Ranger service.
And sets 'default.policy.users' to create default policies that grants
super privileges for 'kudu' user.

Tested with real clusters with both Kerberos and SSL enabled (currently
available at hahao-secure-1.vpc.cloudera.com).

Change-Id: Ic68024e4e4aa8cc59f88134bd18072c391216bd4