CDPD-9417 Remove Non HTTP parcel location - http://archive.cloudera.com/kudu/parcels/latest/
Based on the min/max versions of the 5.15 CSD[1],
the 5.15 CSD is the one used to manage a CDH5.12
Kudu service, so I think we want to keep the
field there and in prior versions.

Later than that (i.e. including kudu-csd-{5.16,
6.0,6.1,6.2}), we don't support parcels, so
we'll want to remove the field on those CSDs too.

[1] https://github.infra.cloudera.com/CDH/kudu/blob/cdpd-master/java/kudu-csd-5.15/src/descriptor/service.sdl

Change-Id: Ia1505f89673a43890b18df00737c93156db444de
Signed-off-by: Kevin Risden <krisden@cloudera.com>