OPSAPS-55657: eliminate discrepancy between 7.0.3 CSD and 7.0-maint
The 7.0.3 CSD is out of sync with the CSD in 7.0-maint. This patch makes
them identical by reverting the following changes:
- the addition of the metrics URL parameters
  - see https://gerrit.sjc.cloudera.com/c/cdh/kudu/+/62590
- the addition of the host config (used by monitoring)
  - see https://gerrit.sjc.cloudera.com/c/cdh/kudu/+/64886
- the addition of the tablet_history_max_age_sec field
  - see https://gerrit.sjc.cloudera.com/c/cdh/kudu/+/62556

There is one difference that remains between the two: the removal of the
repoUrl field (see https://gerrit.sjc.cloudera.com/c/cdh/kudu/+/75252).
AFAICT, this should not lead to staleness, given it has no effect on
gflags or monitoring properties.

Change-Id: I0cfe5d51e20d51e5de56e6d86dc342ad54b84d73