[consensus] small cleanup on Peer::SignalRequest()
It's possible not to hold 'peer_lock_' when scheduling SendNextRequest()
on the thread pool: if 'request_pending_' or 'closed_' changes, it's
handled by Peer::SendNextRequest() as necessary.  It might be an extra
run a thread in Raft pool in the worst case, but it seems better than
blocking the thread running PrepareTask().  Blocking the latter involves
blocking other tasks on corresponding locks and slowing the replication
of a Raft operation, which leads to lower write throughput.

The motivation for this change was seeing the following in the logs
during intensive ingest workloads:

  0531 13:09:40.403224 (+515724us) spinlock_profiling.cc:244] Waited 514 ms on lock 0xa1581150

with the stack symbolized like below:

  kudu::(anonymous namespace)::SubmitSpinLockProfileData(void const*, long)
    src/kudu/util/spinlock_profiling.cc:230
  kudu::consensus::Peer::SignalRequest(bool)
    src/kudu/gutil/spinlock.h:106
  kudu::consensus::PeerManager::SignalRequest(bool)
    src/kudu/consensus/peer_manager.cc:98
  kudu::consensus::RaftConsensus::Replicate(scoped_refptr<kudu::consensus::ConsensusRound> const&)
    src/kudu/consensus/raft_consensus.cc:740 (discriminator 2)
  kudu::tablet::TransactionDriver::Prepare()
    src/kudu/tablet/transactions/transaction_driver.cc:333
  kudu::tablet::TransactionDriver::PrepareTask()
    src/kudu/tablet/transactions/transaction_driver.cc:243
  kudu::ThreadPool::DispatchThread()
    src/kudu/util/threadpool.cc:686

The stack trace pointed to 'peer_lock_' in Peer::SignalRequest(),
consensus_peer.cc,167 as in one of its previous revisions, as the
rest of the stack.

I also did other minor cleanup on the code around.

Change-Id: I899aec0964bc404e6df3742bd48d0d049e52d900
Reviewed-on: http://gerrit.cloudera.org:8080/16021
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>
(cherry picked from commit 710bd6efaa796db4b95ab2c9a13b53abcf5174a2)
  Conflicts:
    src/kudu/tablet/transactions/transaction_driver.cc
    src/kudu/tablet/transactions/transaction_tracker.cc