[consensus] use move semantics for LogEntryBatchPB
As it turned out, in the context of the Log class and related utilities,
it's possible to get rid of the heap allocation and unique_ptr-wrapping
for LogEntryBatchPB.  With this patch, instances of LogEntryBatchPB are
allocated on the stack and passed around using the move semantics.

There aren't "low-level unit" tests for the affected code paths and I'm
not sure it makes sense to introduce one.  However, there is scenario
named 'MultiThreadedLogTest.TestAppends' in the "synthetic" mt-log-test.
To assess performance impact of this patch, I ran the following for
RELEASE builds with and without this patch:

  ./bin/mt-log-test --num_writer_threads=1 --num_batches_per_thread=500000 --num_reader_threads=0 --num_ops_per_batch_avg=1 --verify_log=false --log_segment_size_mb=4096 --never_fsync --unlock_unsafe_flags --gtest_filter='MultiThreadedLogTest.TestAppends' --gtest_repeat=100 | \
  grep 'Time spent inserting 500000 batches' | awk '{print $14}' | \
  sed 's/s$//' | awk 'BEGIN { sum = 0 } { sum += $1 } END { print sum }'

With each run producing summary like below:
  Time spent inserting 500000 batches(1 threads, 500000 per-thread): real 7.506s      user 0.000s     sys 0.000s

I got the following results:
  with the patch:
    711.679

  without the patch:
    736.141

So, the performance at least didn't degrade.

Change-Id: Ib5d3c384dd6f17e8c4c71eec074af9e98827262b
Reviewed-on: http://gerrit.cloudera.org:8080/16075
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>
(cherry picked from commit 8c633e2eb20fca9f45882b291069000c6451c97a)