[csd] CDPD-18222 Add keystore type to config
Ranger client uses a truststore to verify the Ranger admin server's
certificate if TLS is enabled. By default this is a JCE keystore, but
it's not secure enough for FIPS 140-2, so in FIPS clusters BCF keystores
are used. This keystore format is completely different and Ranger client
needs to be configured to use it. The keystore format used by CM is
exposed as keystore_type since OPSAPS-58306.

Change-Id: I51d65a3b231607b1c6705710490642feb976db5d