CDPD-18639 Use defensive locking in FIPS mode
f9f3189a6dbe0636d578d80b1d8e60cf7b2e6686 introduces defensive locking in
some functions that use OpenSSL signature verification that was
necessary to a potential thread-safety bug in SafeLogic CryptoComply for
Servers fork of OpenSSL. This additional locking comes with a
performance penalty in highly concurrent cases unfortunately, so we
decided not to enable it by default, and hide it behind a flag. In CDP,
however, when FIPS approved mode is enabled, it's likely that SafeLogic
CCS is used, so the defensive locking needs to be enabled by default.
This commit changes the default value of the openssl_defensive_locking
flag to true, but it also changes the behavior to additionally check for
FIPS mode, so it can be disabled by setting the flag to false.

Change-Id: Ibbb0d0b8123c25fd78f277f2cb8a94e94db9bdc7
(cherry picked from commit d6a7618b6b67612d102c7ecb4d86538a2264ad5a)