[build] Add /usr/sbin to the PATH in pre_commit_hook.sh
Pre-commit builds were failing with `lsof: command not found` in the
`client_examples-test` even though `lsof` was installed. `lsof` was not found
because the script is not run as root and therefore `/usr/sbin` is not on the
`PATH` by default.

Change-Id: I191db871bacf4b0a04a890c83d1541a23bd2cacb
(cherry picked from commit 6e96f585b608db266dd3051b2a2b4ef7d63ccb99)