CDPD-21419 Fix keystore format
JCE keystores are not secure and they shouldn't be used, especially
in FIPS clusters. CM exposes a config to use BouncyCastle keystores
(BCFKS), which was adhered to by Kudu everywhere, except for one file:
the master TLS truststore, which is currently used by Ranger, is
password-protected and the password is stored in a keystore. This
keystore was hard-coded to be JCEKS. This Commit changes the script to
generate this keystore in the configured format.

The truststore itself is still JCEKS, but that seems to be the case
CDP-wide for all truststores.

Change-Id: I09bde55f071dca82de631b7829b6bacb1d6fe24b