[hms] CDPD-24685: remove incompatible type config requirement
This flips the default behavior to avoid the check for the
hive.metastore.disallow.incompatible.col.type.changes configuration,
since the default changed between CDH and CDP.

This allows us to enable the Kudu-HMS integration in CDP without
detracting from the robustness of Hive. This doesn't fix the underlying
issue with Hive (CDPD-24578, HIVE-24987) that led to Kudu requiring the
config in the first place.

Change-Id: Ia84111deee5fc4bf7252f7b38350785f99dadbda