CDPD-31678: Backport of KUDU-3295
KUDU-3294 Subject line: Fix the number of replicas does not equal the
number of servers

When the tserver is shutdown and its DNS cannot be resolved, it is still
possible for the client to receive the replicas including this tserver,
as a result, discoverTablets will see the servers size not equal to
replica size because the replica does not verify server connectivity.

This patch fixes this by ignoring replica whose server cannot be
resolved.

Change-Id: I7d2e7d49d234e41e66e9c367e061fb7da32e9b47
Reviewed-on: http://gerrit.cloudera.org:8080/17971
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
(cherry picked from commit 94159e5c6713834d86bb89ecb681224cf6eeaf01)