CDPD-49743: backport KUDU-3354: Deflake DisableWriteWhenExceedingQuotaTest
I noticed that the TestDisableWritePrivilegeWhenExceedingSizeQuota
test failed some times in latest code review items because of test
timeout. From the test log we can see maintenance manager tends to
schedule a high score compact operation of some rowsets but do
nothing in the end because 'Compaction resulted in no output rows
(all input rows were GCed!)'. This leads to failures to schedule
flush operations and to hit the size limit until run out of time.

I think this weird high perf score most likely has something to do
with not explicitly initializing the 'value_' of RowsetInfo after
some debugging. The 'value_' can't be updated in FinalizeCDFVector()
either, because in the test scenario it has nearly one key in a
rowset, and the 'total_width' of the rowset is 0.

I run the test 10 times and see no failure with this fix, compared
to 5 failures without the fix but with a DCHECK[1] that indicates
the compaction error occurred in the test case.

[1] https://gerrit.cloudera.org/c/18584/1/src/kudu/tablet/tablet.cc#2135

Change-Id: Id257a1ce04fb2043d79641f770d03393ec5b844b
Reviewed-on: http://gerrit.cloudera.org:8080/18584
Reviewed-by: Yingchun Lai <acelyc1112009@gmail.com>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 687bef89764640746deb27b3279e936acae4446e)
(cherry picked from commit 52653135206050c7d47213ba930db320ef6ced50)