CDPD-6141: Support ranges with custom hash schemas
KUDU-2671 update partition schema in catalog when adding range

When adding a range with custom hash schema to a table, it's necessary
to update the partition schema information stored in the system catalog
correspondingly.  That was missing in one of the previous patches and
this patch addresses the issue.

This patch also adds a test scenario to spot regressions, if any.  The
scenario was failing before the update in CatalogManager introduced
in this patch.  I also addressed nits pointed to by the TidyBot.

This is a follow-up to 250eb90bc0e1f4f472f44de8a23ce213595d5ee7.

Change-Id: I869458fb8bcb06801b54f2b4869e7826322563e0
Reviewed-on: http://gerrit.cloudera.org:8080/18615
Tested-by: Kudu Jenkins
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>
(cherry picked from commit 6909ee4f800da192b72e59680916e5004527b6db)