Add option to disable C++ tests in build script
I tend to use build-and-test.sh during local development. The script
builds and tests the C++ no matter what, and the Java and Python parts
are configurable. Although during Java/Python development the C++ build
is necessary, running all the C++ tests could be omitted. This patch
adds the option to disable running the C++ tests.

For future reference these are the commands to run Java/Python only:

Change-Id: Id3f868c8035876adc3c55bdd2f191afb7197bcbd