CDPD-53605,PVCRB-567,CDPD-54928: [server] KUDU-1945 Auto_incrementing column UPSERT support
Add UPSERT operation support on the server side when the entire row
is present in the request including the auto-incrementing column.
There will be followup patches to update the clients and test the
corresponding changes as well.

With this patch the expected behavior for different operations
from Kudu clients is as follows:

Operation             Presence of auto-incrementing column
INSERT/INSERT_IGNORE  Should not be specified
UPDATE/UPDATE_IGNORE  Has to be specified
UPSERT/UPSERT_IGNORE  Has to be specified
DELETE/DELETE_IGNORE  Has to be specified

This implementation of UPSERT operation behavior in this patch
is to facilitate incremental restores consistently when using
Kudu backup and restore tool.

More details in the design document:
https://docs.google.com/document/d/1-x6F5TkRYMqt2umDBQ5x-Aa4dqKwFYpbxo1A6QoMafo/edit#heading=h.vnagsa4oxwrf

Change-Id: Ib5cc4d80f77c165452572948f68c76fc70394d47
Reviewed-on: http://gerrit.cloudera.org:8080/20061
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>