[server] KUDU-1945 Auto_incrementing column UPSERT support
Add UPSERT operation support on the server side when the entire row
is present in the request including the auto-incrementing column.
There will be followup patches to update the clients and test the
corresponding changes as well.

After this patch below is the different operation support for
tables with auto-incrementing columns from the Kudu client.

Operation             Presence of auto-incrmeneting column
INSERT/INSERT_IGNORE  Should not be specified
UPDATE/UPDATE_IGNORE  Has to be specified
UPSERT/UPSERT_IGNORE  Has to be specified
DELETE/DELETE_IGNORE  Has to be specified

This implemention of UPSERT operation behavior in this patch
is to facilitate incremental restores consistently when using
Kudu backup and restore tool.

More details in the design document:
https://docs.google.com/document/d/1-x6F5TkRYMqt2umDBQ5x-Aa4dqKwFYpbxo1A6QoMafo/edit#heading=h.vnagsa4oxwrf

Change-Id: Ib5cc4d80f77c165452572948f68c76fc70394d47