unset *_proxy env variables for Python3 tests
Commit 459207ff5c103e7d4d60d974504eb8633e188d15 did unset all proxies in
the Jenkins build & test script. It does so in the Python2 part of the
script. Usually we test for both Python2 and Python3, thus the effect of
the unsets in the Python2 part carry over to the Pytho3 section.
However, if someone runs the build script with only Python3 the
proxies are not unset. This patch unsets all proxies in the Python3 part
as well.

Background: I tend to use build-and-test.sh for development because of
convenience reasons. I was working on RHEL9 nodes, where Python2 is not
present. Having the weird network config, I stumbled upon the original
403 error code issue, as build-and-test.sh with only Python3 enabled did
not unset all the proxies.

Change-Id: If23d310ca0cf9cb9dbda60f32a3f675297ffc28e