[Python] KUDU-3351 Add write op metrics
This is a follow-up patch for commit:
0ddcaaabc97c85a4715ae79ff5604feb9b342779, adding per-session write op
metrics to the Python client.

In the test function "test_insert_and_mutate_rows" I added verification
function calls, to check that the metrics are gathered properly.
Only "upsert_ignore_errors" is left out, as UPSERT IGNORE is not yet
supported by the Python client. I'm planning to address that in the next
patch.

Change-Id: Id76f8f0cb11ef5e4b9d06508a39492bc6b0109a9