Add code coverage support for Java
This change adds JaCoCo-based code coverage support for the Java modules
in the project. The coverage data is collected during test runs and can
be viewed in HTML format using the `./gradlew jacocoAggregatedReport`
task.

With this setup, it's now possible to generate and publish both C++ and
Java coverage reports from a single Jenkins job.

As demonstrated on https://jenkins.kudu.apache.org/job/coverage/,
there's no need to configure and maintain separate jobs for each
language.

Additionally, a small script runs `jacocoLogAggregatedCoverage` and
appends the aggregated JaCoCo metrics (e.g., summary line coverage) to
the bottom of the generated report. This step is necessary because
JaCoCo does not include these summary values in the HTML report by
default.

Change-Id: I15816adb14785eade28581f75918f6848457d190