thirdparty: FlatBuffers build fix
Add FlatBuffers to the explicit dependency list in
thirdparty/build-thirdparty.sh. This allows running:

  ./thirdparty/build-thirdparty.sh flatbuffers

Without this, the script fails with: "Unknown module: flatbuffers".

When adding a new package, I often run download-thirdparty.sh and then
build-thirdparty.sh with the package name to rebuild just that package,
avoiding a full rebuild. That's when I hit the error above.

Change-Id: I9bf6d9de9413dd914d21e50b81ccb80265a14b1c